[BE][feat] : 댓글 API 입력값 검증 - params의 id가 유효한 id인지 확인

+++ backend/services/comment/deleteComment.js
@@ -3,6 +3,8 @@ const { comment } = require('../../sequelize/models');
 const deleteComment = async (req, res) => {
   try {
     const { id } = req.params;
+    const parsedId = parseInt(id, 10);
+    if (Number.isNaN(parseInt) || parsedId < 1) throw new TypeError();
 
     const result = await comment.destroy({
       where: {
